working on rsw_cstr and other tweaks

+++ c_utils.c
@@ -382,7 +382,7 @@ char* readstring(c_array* input, char delim, size_t max_len)
 
 
 
-c_array slice_c_array(c_array array, int start, int end)
+c_array slice_c_array(c_array array, long start, long end)
 {
 	c_array a = { NULL, array.elem_size, 0 };
 	int len;
@@ -553,7 +553,7 @@ char* trim(char* str)
 }
 
 
-long find(c_array haystack, c_array needle)
+size_t find(c_array haystack, c_array needle)
 {
 	byte* result = haystack.data;
 	byte* end = haystack.data + haystack.len*haystack.elem_size;
@@ -565,7 +565,7 @@ long find(c_array haystack, c_array needle)
 		}
 	}
 
-	return -1;
+	return -1; /* make a macro or static const size_t npos = -1 ? */
 }
 
 /*
